rust: added a loader, a Cargo.toml and an example.
